dividend usable en multiple
Después de muchos cambios creo que esta rama ha servido su propósito.

-d se podrá usar con -j en multiple sin que haya colisión.

También he creado otro tipo de regla que representa a las que no generan ninguna útil por su cuenta y necesita que se usen las de sus factores primos.

Además -d usa BigInteger para no limitar las capacidades de la aplicación arbitrariamente.

+++ Operaciones/ReglaCero.cs
@@ -1,4 +1,5 @@
 ﻿using Operaciones.Recursos;
+using System.Numerics;
 using System.Text.Json.Serialization;
 
 namespace Operaciones {
@@ -32,7 +33,7 @@ namespace Operaciones {
 		[JsonPropertyName("type")]
 		public CasosDivisibilidad Tipo => CasosDivisibilidad.DIVISOR_ZERO;
 
-		public string AplicarRegla(long dividendo) => TextoCalculos.ReglaExplicadaCero;
+		public string AplicarRegla(BigInteger dividendo) => TextoCalculos.ReglaExplicadaCero;
 		public override string ToString() {
 			return ReglaExplicada;
 		}
